+++ qbitwave/__init__.py
@@ -1,10 +1,14 @@
 from .qbitwave import QBitwave
+from .qbitspinor import QBitSpinor
 from .qbitwave_mdl import QBitwaveMDL
 from .qbitwavend import QBitwaveND
+from .spectral_field_mdl import SpectralFieldMDL
 
 
 __all__ = [
     "QBitwave",
+    "QBitSpinor",
     "QBitwaveMDL",
-    "QBitwaveND"
+    "QBitwaveND",
+    "SpectralFieldMDL",
 ]